refactor: localStorage util 함수 빈 값 리턴 방어 및 추상화
