+++ docs/REQUIREMENTS.md
@@ -41,8 +41,8 @@
 - [x] `updateRestaurantList()` 는 `render` 로 분리하기
 - [x] `updateRestaurantList()` 에서 local storage 저장 함께 하기
 - [x] template literal 포맷 맞추기
-- [ ] `getAllDataOnLocalStorage()` 안전하게 수정하기
-- [ ] `getAllDataOnLocalStorage()` 추상화하거나 domain으로 변경하기
+- [x] `getAllDataOnLocalStorage()` 안전하게 수정하기
+- [x] `getAllDataOnLocalStorage()` 추상화하거나 domain으로 변경하기
 - [ ] 이벤트 핸들러 유틸 추상화 심화하기
 - [ ] `executeChangeEventListener()` 네이밍 수정하기
 - [x] `export interface CustomError extends Error {}` 제거하기
