--- src/constants/localStorage.ts
@@ -1,1 +1,0 @@
-export const LOCAL_STORAGE_KEY: string = 'restaurant';
