+++ src/constants/values.ts
@@ -2,3 +2,5 @@ export const NAME_LENGTH = {
   MIN: 1,
   MAX: 15,
 };
+
+export const LOCAL_STORAGE_KEY: string = 'restaurant';
